requirements.txt
python-telegram-bot
requests
rapidfuzz
ujson
